test: add API integration tests + fix 4 bugs found via TDD
Tests (32 tests, 107 assertions, all passing):
- PlayersControllerTest: 9 tests (auth, CRUD, user scoping)
- GameTypesControllerTest: 6 tests (CRUD, validation, uniqueness)
- GamesControllerTest: 8 tests (CRUD, game completion, rank calculation)
- RoundsControllerTest: 4 tests (auto-numbering, ownership, completed game)
- ScoresControllerTest: 5 tests (scoring, total recalculation, bulk ops)
- ApiTestCase: shared base with auth helpers, DB cleanup, factory methods

Bug fixes discovered through testing:
- routes.php: :id -> {id} (CakePHP 5 route parameter syntax)
- Application.php: JWT config 'secret' -> 'secretKey'
- JwtAuthTrait: handle JWT payload identity (load User entity from sub claim)
- RoundsController: cast auto-number to int (DB aggregate returns string)

+++ api/src/Controller/Api/RoundsController.php
@@ -176,7 +176,7 @@ class RoundsController extends AppController
                 ->where(['game_id' => $data['game_id']])
                 ->select(['max_round' => $this->Rounds->find()->func()->max('round_number')])
                 ->first();
-            $data['round_number'] = ($maxRound->max_round ?? 0) + 1;
+            $data['round_number'] = (int)(($maxRound->max_round ?? 0) + 1);
         }
 
         $round = $this->Rounds->newEmptyEntity();
